% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometrics.R
\name{calc_aspirated_wb}
\alias{calc_aspirated_wb}
\title{Calculate psychrometric wet bulb temperature}
\usage{
calc_aspirated_wb(temp, RH)
}
\arguments{
\item{temp}{Air temperature in degrees Celsius.}

\item{RH}{Relative humidity in percent.}
}
\value{
Psychrometric (aspirated) wet bulb temperature in degrees Celsius.
}
\description{
Uses the Stull (2011) approximation for aspirated (psychrometric)
wet bulb temperature.
}
\references{
Stull, R. (2011) Wet-Bulb Temperature from Relative Humidity and Air
Temperature. Journal of Applied Meteorology and Climatology, 50(11),
2267-2269.
}
\keyword{internal}
