% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometrics.R
\name{calc_dew_point}
\alias{calc_dew_point}
\title{Calculate dew point temperature}
\usage{
calc_dew_point(temp, RH)
}
\arguments{
\item{temp}{Air temperature in degrees Celsius.}

\item{RH}{Relative humidity in percent.}
}
\value{
Dew point temperature in degrees Celsius.
}
\description{
Uses the inverse of the Buck equation to derive dew point from
temperature and relative humidity.
}
\keyword{internal}
