% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometrics.R
\name{calc_sat_vp}
\alias{calc_sat_vp}
\title{Calculate saturation vapour pressure using the Buck equation}
\usage{
calc_sat_vp(temp)
}
\arguments{
\item{temp}{Temperature in degrees Celsius.}
}
\value{
Saturation vapour pressure in kPa.
}
\description{
Calculate saturation vapour pressure using the Buck equation
}
\keyword{internal}
