% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_temp.R
\name{calculate_globe_temp}
\alias{calculate_globe_temp}
\title{Calculate globe temperature (vectorised)}
\usage{
calculate_globe_temp(
  temp,
  wind_speed,
  direct_solar,
  diffuse_solar,
  zenith,
  albedo = 0.12,
  verbose = FALSE
)
}
\arguments{
\item{temp}{Air temperature in degrees Celsius.}

\item{wind_speed}{Wind speed in m/s.}

\item{direct_solar}{Direct beam solar radiation in W/m^2.}

\item{diffuse_solar}{Diffuse sky solar radiation in W/m^2.}

\item{zenith}{Solar zenith angle in degrees.}

\item{albedo}{Ground albedo (0--1). Default 0.12 for asphalt.}

\item{verbose}{If \code{TRUE}, pass observation index to solver for warnings.}
}
\value{
Numeric vector of globe temperatures in degrees Celsius.
}
\description{
Vectorised wrapper around \code{solve_globe_temp()} using \code{\link[purrr:pmap]{purrr::pmap_dbl()}}.
}
\keyword{internal}
