% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_wet_bulb.R
\name{calculate_natural_wet_bulb}
\alias{calculate_natural_wet_bulb}
\title{Calculate natural wet bulb temperature (vectorised)}
\usage{
calculate_natural_wet_bulb(
  temp,
  RH,
  pressure,
  wind_speed,
  globe_temp,
  verbose = FALSE,
  show_progress = FALSE
)
}
\arguments{
\item{temp}{Air temperature in degrees Celsius.}

\item{RH}{Relative humidity in percent.}

\item{pressure}{Atmospheric pressure in kPa.}

\item{wind_speed}{Wind speed in m/s.}

\item{globe_temp}{Globe temperature in degrees Celsius.}

\item{verbose}{If \code{TRUE}, show warnings.}

\item{show_progress}{If \code{TRUE}, display a progress bar for large inputs.}
}
\value{
Numeric vector of natural wet bulb temperatures in degrees Celsius.
}
\description{
Vectorised wrapper around \code{solve_natural_wb_single()} using
\code{\link[purrr:pmap]{purrr::pmap_dbl()}}, with optional progress bar.
}
\keyword{internal}
