% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{clus_save}
\alias{clus_save}
\title{Output samples}
\format{A list of data.tables}
\source{
output of TWLsample function
}
\usage{
data(data_and_output)
}
\description{
5000 iterations from output of TWLsample function
}
\examples{
data(data_and_output)
ls()
}
\keyword{datasets}
