% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{misaligned}
\alias{misaligned}
\title{Progressively misaligned cluster annotation}
\format{A list of data.tables}
\source{
simulated
}
\usage{
data(data_and_output)
}
\description{
Example annotation information for simulated 
data of progressively misaligned clusters
}
\examples{
data(data_and_output)
ls()
}
\keyword{datasets}
