% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{misaligned_mat}
\alias{misaligned_mat}
\title{Progressively misaligned cluster data matrices}
\format{A list of matrices}
\source{
simulated
}
\usage{
data(data_and_output)
}
\description{
Simulated data of progressively misaligned clusters
on which to fit a TWL model.
}
\examples{
data(data_and_output)
ls()
}
\keyword{datasets}
