% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{outpu_new}
\alias{outpu_new}
\title{Output PSMs}
\format{A list of matrices}
\source{
output of pairwise_clus function
}
\usage{
data(data_and_output)
}
\description{
Posterior similar matrices, output of pairwise_clus function
}
\examples{
data(data_and_output)
ls()
}
\keyword{datasets}
