% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_wet_bulb.R
\name{solve_natural_wb_single}
\alias{solve_natural_wb_single}
\title{Solve natural wet bulb temperature for a single observation}
\usage{
solve_natural_wb_single(
  temp,
  RH,
  pressure,
  wind_speed,
  globe_temp,
  index = NULL
)
}
\arguments{
\item{temp}{Air temperature in degrees Celsius.}

\item{RH}{Relative humidity in percent.}

\item{pressure}{Atmospheric pressure in kPa.}

\item{wind_speed}{Wind speed in m/s.}

\item{globe_temp}{Globe temperature in degrees Celsius.}

\item{index}{Observation index for warnings (optional).}
}
\value{
Natural wet bulb temperature in degrees Celsius.
}
\description{
Iteratively solves the heat balance on an unventilated wet bulb
thermometer using Newton's method, based on the Brake (2001) methodology.
}
\keyword{internal}
