% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_twl.R
\name{solve_twl_single}
\alias{solve_twl_single}
\title{Solve TWL for a single observation}
\usage{
solve_twl_single(
  temp,
  wind_speed,
  RH,
  direct_solar,
  diffuse_solar,
  pressure,
  solar_radiation,
  pa,
  temp_dewpoint,
  wet_bulb,
  globe_temp,
  trad,
  hc,
  max_core_temp,
  max_sweat_rate,
  Icl,
  icl,
  LR,
  lambda,
  fr,
  index = NULL
)
}
\arguments{
\item{temp}{Air temperature in degrees Celsius.}

\item{wind_speed}{Wind speed in m/s.}

\item{RH}{Relative humidity in percent.}

\item{direct_solar}{Direct solar radiation in W/m^2.}

\item{diffuse_solar}{Diffuse solar radiation in W/m^2.}

\item{pressure}{Atmospheric pressure in kPa.}

\item{solar_radiation}{Total solar radiation in W/m^2.}

\item{pa}{Actual vapour pressure in kPa.}

\item{temp_dewpoint}{Dew point temperature in degrees Celsius.}

\item{wet_bulb}{Natural wet bulb temperature in degrees Celsius.}

\item{globe_temp}{Globe temperature in degrees Celsius.}

\item{trad}{Mean radiant temperature in degrees Celsius.}

\item{hc}{Convective heat transfer coefficient in W/(m^2.K).}

\item{max_core_temp}{Maximum core temperature in degrees Celsius.}

\item{max_sweat_rate}{Maximum sweat rate in kg/(m^2.hr).}

\item{Icl}{Clothing insulation in clo.}

\item{icl}{Clothing vapour permeability (0--1).}

\item{LR}{Lewis relation in K/kPa.}

\item{lambda}{Latent heat of evaporation in kJ/kg.}

\item{fr}{Posture factor.}

\item{index}{Observation index for warnings (optional).}
}
\value{
TWL in W/m^2.
}
\description{
Iteratively solves the heat balance equations to find the maximum
sustainable metabolic rate using a three-zone evaporation model and
Newton's method.
}
\keyword{internal}
