\name{twl-package}
\alias{twl-package}
\alias{twl}
\docType{package}
\title{
\packageTitle{twl}
}
\description{
\packageDescription{twl}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{twl}
\packageIndices{twl}
}
\author{
\packageAuthor{twl}

Maintainer: \packageMaintainer{twl}
}
\references{Swanson DM, Lien T, Bergholtz H, Sorlie T, Frigessi A, Investigating Coordinated Architectures Across Clusters in Integrative Studies: a Bayesian Two-Way Latent Structure Model, 2018, doi: 10.1101/387076, Cold Spring Harbor Laboratory, https://www.biorxiv.org/content/early/2018/08/07/387076.full.pdf.}
\keyword{ package }
