% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TWL_CONSTANTS}
\alias{TWL_CONSTANTS}
\title{Physical and physiological constants for TWL calculation}
\format{
An object of class \code{list} of length 8.
}
\usage{
TWL_CONSTANTS
}
\description{
A list of constants used throughout the TWL calculations:
\describe{
\item{STEFAN_BOLTZMANN}{Stefan-Boltzmann constant, 5.67e-8 W/(m^2.K^4)}
\item{ABS_ZERO}{Absolute zero offset, 273.15 K}
\item{LATENT_HEAT_EVAP}{Latent heat of evaporation at skin temp, 2430 kJ/kg}
\item{VIEW_EMISSIVITY_FACTOR}{Combined view (0.8) and emissivity (0.95) factor, 0.76}
\item{BULB_DIA}{Standard wick diameter, 0.004 m}
\item{AIR_THERMAL_CONDUCTIVITY}{Thermal conductivity of air, 0.028 W/(m.K)}
\item{GLOBE_DIAMETER}{Black globe diameter, 0.15 m}
\item{GLOBE_EMISSIVITY}{Globe emissivity, 0.95}
}
}
\keyword{internal}
