% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWL-package.R
\docType{package}
\name{TWL-package}
\alias{TWL}
\alias{TWL-package}
\title{TWL: Thermal Work Limit Calculation}
\description{
Calculates the Thermal Work Limit (TWL), the maximum sustainable metabolic rate (W/m²) for workers in thermally stressful environments. Based on the methodology of Brake & Bates (2002) \doi{10.1080/104732202317070947}.
}
\author{
\strong{Maintainer}: RJS \email{author@example.com}

}
\keyword{internal}
